Two file updates-2.3.2
green stocks spreadsheet has been update to contain an output table and an array that will be used in a for loop. 
The practice sheet contains the skill drill in module 2.3.2 where I practiced nested for loops to fill cells and then researched how to use a for loop to clear those cells.