utilizing visual formatting with VBA-2.4.2
the "practiceforloops" sheet contains code where I created a checkerboard using nested for loops and conditionals. 
The new green stock sheet is updated to contain visual formatting aid where positive yearly returns for each stock are green and negative are red and the numbers are further formatted to ease reading.